Fix build

+++ src/Processors/OffsetTransform.h
@@ -33,7 +33,7 @@ private:
     size_t num_finished_port_pairs = 0;
 
 public:
-    OffsetTransform(const Block & header_, size_t offset_, size_t num_streams = 1);
+    OffsetTransform(const Block & header_, UInt64 offset_, size_t num_streams = 1);
 
     String getName() const override { return "Offset"; }
